separated out the backplaneadapter code and backplane code to aid reading
